requirements.txt
qrcode>=7.4.2
openpyxl>=3.1.2
waitress>=3.0.0
flask>=3.0.3
pillow>=10.3.0
